+++ tests/testthat/testutils.rmo/R/bernstein-fuzzing.R
@@ -334,3 +334,40 @@ setMethod(
         bf
     }
 )
+
+#' @rdname fuzzy_bf-methods
+#' @aliases fuzzy_bf,ConvexCombinationOfBernsteinFunctions,ANY-method
+#'
+#' @seealso [rmo::ConvexCombinationOfBernsteinFunctions-class]
+#' @examples
+#' bf <- fuzzy_bf(rmo::ConvexCombinationOfBernsteinFunctions())
+#' bf <- fuzzy_bf(rmo::ConvexCombinationOfBernsteinFunctions(
+#'     coefficients = rep(1, 4L),
+#'     points = list(
+#'         rmo::LinearBernsteinFunction(),
+#'         rmo::ConstantBernsteinFunction(),
+#'         rmo::AlphaStableBernsteinFunction(),
+#'         rmo::GammaBernsteinFunction()
+#'     )
+#' ))
+#'
+#' @importFrom stats rpois
+#' @importFrom methods setMethod validObject is
+#' @importFrom rmo ConvexCombinationOfBernsteinFunctions
+#' @export
+setMethod(
+    "fuzzy_bf", "ConvexCombinationOfBernsteinFunctions",
+    function(bf) {
+        if (isFALSE(length(bf@points) > 1L)) {
+            n <- rpois(n = 1, lambda = 10)
+            bf@points <- sapply(seq_len(n), function(x) fuzzy_bf())
+        } else {
+            bf@points <- sapply(bf@points, fuzzy_bf)
+            n <- length(bf@points)
+        }
+        bf@coefficients <- rexp(n)
+        validObject(bf)
+
+        bf
+    }
+)